# docs/swagger/manager/manager_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing shards and cluster state",
        "title": "Sharding System Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@sharding-system.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/execute": {
            "post": {
                "description": "Executes a SQL query on the shard determined by the shard key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "router"
                ],
                "summary": "Execute a query on a shard",
                "parameters": [
                    {
                        "description": "Query Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.QueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query executed successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.QueryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/shard-for-key": {
            "get": {
                "description": "Returns the shard ID that handles the given key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "router"
                ],
                "summary": "Get shard ID for a key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shard key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shard ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_sharding-system_pkg_models.QueryRequest": {
            "type": "object",
            "properties": {
                "consistency": {
                    "description": "\"strong\" or \"eventual\"",
                    "type": "string"
                },
                "options": {
                    "type": "object",
                    "additionalProperties": true
                },
                "params": {
                    "type": "array",
                    "items": {}
                },
                "query": {
                    "type": "string"
                },
                "shard_key": {
                    "type": "string"
                }
            }
        },
        "github_com_sharding-system_pkg_models.QueryResponse": {
            "type": "object",
            "properties": {
                "latency_ms": {
                    "type": "number"
                },
                "row_count": {
                    "type": "integer"
                },
                "rows": {
                    "type": "array",
                    "items": {}
                },
                "shard_id": {
                    "type": "string"
                }
            }
        },
...
    }
}